django/requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
boto3==1.17.7
botocore==1.20.7
celery==4.4.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.4
defusedxml==0.7.0rc2
Django==3.1.6
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-filter==2.4.0
django-storages==1.11.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
flower==0.9.5
fuzzywuzzy==0.18.0
gunicorn==20.0.4
humanize==3.2.0
idna==2.10
jmespath==0.10.0
kombu==4.6.11
oauthlib==3.1.0
pika==1.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-Levenshtein==0.12.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
